augment-projects/passwordz/requirements.txt
pyside6>=6.0.0
pywin32>=300
cryptography>=40.0.0
pyyaml>=6.0
